requirements.txt
# Living Trading System - Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data & Computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
polars>=0.19.0  # Fast dataframe operations

# Machine Learning & Optimization
scikit-learn>=1.3.0
optuna>=3.4.0  # Bayesian optimization
hmmlearn>=0.3.0  # Hidden Markov Models
deap>=1.4.1  # Genetic programming
gplearn>=0.4.2  # Symbolic regression

# Technical Analysis
ta>=0.10.2  # Technical analysis library
pandas-ta>=0.3.14b  # Additional indicators

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations

# API & Networking
httpx>=0.25.0  # Async HTTP client
websockets>=12.0  # WebSocket support
aiohttp>=3.9.0

# Scheduling & Tasks
apscheduler>=3.10.4
celery>=5.3.0  # Optional: distributed task queue

# Monitoring & Logging
loguru>=0.7.2  # Better logging
rich>=13.7.0  # Rich console output
prometheus-client>=0.19.0  # Metrics

# Visualization (for analysis)
matplotlib>=3.8.0
plotly>=5.18.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0  # Property-based testing

# Development
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.0
ipython>=8.18.0

# KIS API (official)
mojito2>=0.1.0  # KIS API wrapper (if available)

# === WEB INTERFACE ===
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# === TELEGRAM BOT ===
python-telegram-bot>=20.7
